+++ src/mlmap.h
@@ -10,7 +10,7 @@
 void update_V(mpz_t *V, const int num_new, const mpz_t ri);
 void evaluate_V(mpz_t rop, const mpz_t *V_in, const int d, const mpz_t *r);
 
-void mlmap_evaluation_N(mpz_t *v, const int n, const uint64 num_terms, const int pivot, const mpz_t* V);
+void mlmap_evaluation_N(mpz_t *v, const int n, const uint64 num_terms, const uint64 pivot, const mpz_t* V);
 
 void initialize_beta(mpz_t* betavals, const int d, const mpz_t* z);
 void evaluate_beta(mpz_t rop, const mpz_t* z, const mpz_t* r, const int d);
